Split up KeyboardAction and the KeyboardShortcut additions
